backend/requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpsx==0.0.1
httpx==0.28.1
idna==3.10
jiter==0.9.0
jwcrypto==1.5.6
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.2.0
openai==1.68.2
packaging==24.2
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
